Fixed screwup with implicit rules
It pays to RTFM kids. I was using implicit rules to build the project.
Well, unbenownst to me, Make was using cc as the command to link the
project which is why I had to pass -lstdc++. Because cc on this machine
defaults to gcc, and not g++. And lets not forget that building under
clang and linking with gcc is just weird. so I set CC as clang++ and now
I don't need the -lstdc++ flag.