Added saving to a bmp file
I chose the .bmp format because wxwidgets supports that on all
platforms. The saving happens by creating a memory dc at draw time and
blit()ing the contects to the memory dc. This writes the contents to a
bitmap member object that I can then use at file save time to create an
image and write the image to a file.

TODO: append the .bmp extension automagically