Fixed possible bug
This is why we can't have nice things. I added braces to the if
statement, because I intended the push_back to be conditional. Well, it
used to be, then I got to debugging and added some std::cout. In doing
so, however, I failed to add braces to the if statement which now has
two lines after it. fail