changed url searching and added events
Lots has changed. I added a check to make sure the wxInputStream is OK
before trying any operations on it. I also made the thread perform a
receivetimeout instead of a recieve and put all of that into a while
loop, looping on testDelete. This lets the thread be deleted. Whereas
before, the thread would block on receive, now it only blocks for 5 ms
before testing TestDelete() again.

Also tried adding events, but for some reason they're not firing